Transpiler: single Int64 integral tier, decimal-safe division, MEMBER-TYPE evidence
Three interlocking pieces of the integral-typing project:

1. Division fix (unblocks int typing everywhere). Harbour `/` is
   ALWAYS float — 7/2 == 3.5 — but C# int/int truncates. HB_EO_DIV now
   emits `(decimal)(left) / right` when BOTH operands are statically
   C#-integral (hb_csExprIsCsIntegral: integer literals, INTEGER
   locals, integral #defines, ORM def-class integral fields, `as int`
   members, and +,-,*,% arithmetic over those). A decimal on either
   side already divides decimally, so ordinary divisions gain no
   casts. `%` needs nothing (int%int is exact); `^` already routes
   through the decimal HbRuntime.Pow. With reads safe, Pass 2.5's
   division-READ disqualifier is gone — only fractional WRITES
   (fractional assigns, /= %= ^= compounds) force decimal now — and
   W0029 skips provably-integral RHS (int literals were false-firing).
   test84 pins runtime parity across every integral source dividing.

2. Single integral tier: ALL integers are C# long (Alex's ruling —
   Harbour's integral type is 64-bit; an Int32 sub-tier only
   manufactured width distinctions Harbour doesn't have: len-9 vs
   len-10 def fields, int-vs-long #define tiers, AS INTEGER vs AS
   LONG). hb_csTypeMap INTEGER (and legacy `as int`) -> long; every
   emitted cast is now (long) — write coercions, FOR starts, class
   INIT coercions, array subscripts `[(long)(expr) - 1]` and dims
   `new dynamic[(long)(n)]`; zero `(int)` remains in generated code.
   gendefines emits `const long` for every integral value, and both
   integral map tokens infer as INTEGER — so >Int32 bit-flag masks
   are properly integral now instead of decimal fallbacks. The
   annotation keyword is AS INTEGER alone, meaning Int64; it already
   parses in the transpiler AND compiles in real Harbour 3.2 class
   VARs, so the source-side sweep needs no grammar work. test82
   reworded: BIGVAL (2^32) pins that the tier is Int64.

3. MEMBER-TYPE evidence collector (audit-only recon for annotating
   POSStatus/Transaction/TranLine members). A user-class member
   meeting an ORM-typed field in an assignment (either direction) or
   comparison ties the member to the field's exact type — the data
   flow that reaches Transaction's nSale* members, which are renamed
   copies of ORM data no name-join can see. Emitted per site into
   --type-audit as MEMBER-TYPE rows (Self class parsed from the
   `Class::Class__Method` reftab key, now carried in the type env);
   MEMBER-DIV marks fractional compound writes, the one shape that
   still blocks annotation. scripts/member-types.py (easipos-
   transpiled) aggregates into per-member AS INTEGER proposals:
   currently 76 members, zero conflicts.

verify.sh green: 86/86 all stages + 6 negative.

+++ src/transpiler/tests/csout/test13.cs
@@ -23,7 +23,7 @@ public static partial class Program
     public static decimal SumEvenTo(decimal nMax = default)
     {
         decimal nSum = 0;
-        int i = default;
+        long i = default;
 
         for (i = 1; i <= nMax; i++)
         {
@@ -45,7 +45,7 @@ public static partial class Program
     {
         string cItem = default;
         decimal nPos = 0;
-        int i = default;
+        long i = default;
 
         // Nested FOR with EXIT
         for (i = 1; i <= HbRuntime.Len(aItems); i++)
